The message-tray is now updated to show up as a hidden side-nav on the right side of the main page.
The message tray will be shown when user clicks on the "<" arrow on the right side of the page. The message tray can be hidden back by clicing the ">" button.

The message tray shows responses and error messages with a time-stamp and it is sorted by newest data on top.

This change list is only for the UI layout changes.

A new component "sidebadbutton" has been added as has a new service "sidenav.service".

The display in the app.component.html is modified to encapsulate all the elements within a set of "mat-drawer" and "mat-sidenav" containers.

The "<" and ">" mat-icon buttons are placed in a sidenav of their own and displayed using "ngif / else" logic to show and hide the message tray

Signed-off-by: jiyer2 <jiyer2@nmdp.org>